pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fpl-pipeline"
version = "0.1.0"
description = "Fantasy Premier League prediction pipeline with ML and Monte Carlo simulation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    # Data processing
    "polars>=1.0.0",
    "duckdb>=1.0.0",
    "pandas>=2.0.0",

    # ML & Statistics
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "pulp>=2.7.0",
    "numpy>=1.24.0",

    # Data ingestion
    "soccerdata>=0.15.0",
    "requests>=2.31.0",
    "supabase>=2.0.0",

    # Validation
    "pandera>=0.18.0",

    # Tracking
    "mlflow>=2.9.0",

    # UI
    "streamlit>=1.30.0",
    "streamlit-aggrid>=1.0.0",
    "plotly>=5.18.0",
    "matplotlib>=3.7.0",

    # Database
    "psycopg2-binary>=2.9.0",

    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    # Linting & formatting
    "ruff>=0.1.8",
    "black>=23.12.0",
    "mypy>=1.8.0",
    "isort>=5.12.0",
    # Pre-commit
    "pre-commit>=3.6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "N"]
ignore = [
    "N806",  # Variable in function should be lowercase - common in ML code for X, y
    "N999",  # Invalid module name - Streamlit pages use numeric prefixes
    "E501",  # Line too long - Black handles this differently
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing -v"

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
